Event interaction UI: refactor implementation of OnNavigationItemSelectedListener
